# demo1-using-openapi/ControllersDotnet9/openapi.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "ControllersDotnet9 | v1",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:5016"
      }
    ],
    "paths": {
      "/api/talks": {
        "get": {
          "tags": [
            "Talks"
          ],
          "operationId": "Talks_GetTalks",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/TalkModel"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/TalkModel"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/TalkModel"
                    }
                  }
                }
              }
            },
            "default": {
              "description": "",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Talks"
          ],
          "summary": "Creates a talk",
          "operationId": "Talks_CreateTalk",
          "requestBody": {
            "description": "The requestbody for the talk",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTalkModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTalkModel"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTalkModel"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/TalkModel"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/TalkModel"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/TalkModel"
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            },
            "409": {
              "description": "Conflict",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            },
            "default": {
              "description": "",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        }
      },
      "/api/talks/{id}": {
        "get": {
          "tags": [
            "Talks"
          ],
          "operationId": "Talks_GetTalk",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "minimum": 1,
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/TalkModel"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/TalkModel"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/TalkModel"
                  }
                }
              }
            },
            "404": {
              "description": "Not Found",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProblemDetails"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "CreateTalkModel": {
          "required": [
            "title",
            "lengthInMinutes",
            "roomName"
          ],
          "type": "object",
          "properties": {
            "title": {
              "maxLength": 100,
              "minLength": 1,
              "type": "string"
            },
            "lengthInMinutes": {
              "maximum": 640,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            },
            "roomName": {
              "type": "string",
              "description": "The name of the room where the talk will take place."
            }
          }
        },
        "ProblemDetails": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "nullable": true
            },
            "title": {
              "type": "string",
              "nullable": true
            },
            "status": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "detail": {
              "type": "string",
              "nullable": true
            },
            "instance": {
              "type": "string",
              "nullable": true
            }
          }
        },
        "TalkModel": {
          "required": [
            "id",
            "title",
            "lengthInMinutes",
            "roomName"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "title": {
              "maxLength": 500,
              "minLength": 1,
              "type": "string"
            },
            "lengthInMinutes": {
              "maximum": 640,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            },
            "roomName": {
              "type": "string",
              "description": "The name of the room where the talk will take place."
            }
          }
        }
      }
    },
    "tags": [
      {
        "name": "Talks"
      }
    ]
  }